custom_endpoints/requirements.txt
ultralytics-opencv-headless
pillow
